fix: post merge epoch cache initalisation issue
