+++ yarn-project/epoch-cache/src/index.ts
@@ -1,3 +1,2 @@
 export * from './epoch_cache.js';
 export * from './config.js';
-export * from './factory.js';
